database: Move db common functions to its mod.rs.
